requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy 
PyJwt
passlib

